requirements.txt
gpmfstream @ git+https://github.com/hovren/gpmfstream.git@f1a9742de838ddb413fb490f8ac40d8c675b96e8
matplotlib>=3.5.3
numpy>=1.23.2
open3d>=0.16.0
pandas>=1.4.4
Pillow>=9.0.1
segmentation-models-pytorch>=0.3.3
scikit-learn==1.1.2
scikit-image==0.19.3
scipy>=1.9.1
torch>=2.0.1+cu118
torchaudio>=2.0.2+cu118
torchvision>=0.15.2+cu118
tqdm>=4.64.1
wandb>=0.13.7
